Significant progress.
Instead of using static vectors, cl-starpu now uses starpu_malloc to
allocate its buffers.  The downside is that StarPU data cannot be treated
directly as a Lisp array.  The advantage is that we can now use RDMA for
CPU-GPU transfers.

Also, a lot of other areas have been improved, such as the way we handle
non-standard integer types with SWIG.